medication prescripion, hide refresh btn on files and more
