+++ src/components/Patient/diagnosis/list.tsx
@@ -3,8 +3,8 @@ import { format } from "date-fns";
 import { useTranslation } from "react-i18next";
 
 import { TableSkeleton } from "@/components/Common/SkeletonLoading";
-import { EmptyState } from "@/components/Medicine/MedicationRequestTable";
 import { EncounterAccordionLayout } from "@/components/Patient/EncounterAccordionLayout";
+import { EmptyState } from "@/components/Patient/MedicationPrescriptionTab";
 
 import query from "@/Utils/request/query";
 import {
